moving DBPALibraryWrapper into its own .h and .cc files. No tests yet.
